Preload lib and models to allow initializing properties
In order for dm to be able to infer all FK properties
correctly, we have to "touch" them at least once,
*after* all participating models are known to dm.

  ::DataMapper::Model.descendants.each do |model|
    model.relationships.each_value { |r| r.child_key }
  end

The current implementation preloads all content
below Rails.root/lib and Rails.root/app/models.
This is necessary because rails currently doesn't
provide a way to add a hook after the application
models have been loaded, because it loads the models
lazily.