Require rails/generators/* files from generators
This makes the generators show up when running

  ./script/generate

However, they are currently not yet working. Doing

  ./script/generate data_mapper:model

will result in

  Don't know how to build task 'data_mapper:model'

I'm not yet sure why this is, but there also still
seems to be a problem with "namespacing" the
generators. Currently they are put below the Rails
namespace and show up like so in ./script/generate

  Rails:
    ...
    data_mapper:migration
    data_mapper:model
    data_mapper:observer
    ...

Probably, this is due to the current module
organization in rails3_datamapper, which uses the
Rails namespace instead of the DataMapper namespace.
This will change soon.

Status: generators are still not useable.