Skip loading of generators on after_initialize process.
Looks like there's some sort of load order issues
going on here. To see the issue run the following
more than once.

  script/rails generate generator awesome

And the error you get:

  railties/lib/rails/generators/base.rb:214:in `inherited':
  undefined method `subclasses' for Rails::Generators:Module

Skipping the generators here lets Rails continue
its loading and the generators get loaded down the
road.

It looks like not enough of the Rails framework
is loaded at this point to successfully load the
generators, so I just skip them here and let them
get loaded later on. I'm not sure if this is the
best way to do it or not, but it has been working
fine for me.