Proper Gemfile for easy dev environment setup
1)

Issue the following commands to get a fully functional
development environment including datamapper and rails
up and running within a minute

  gem install bundler # if this is new for you
  bundle install
  rake spec

Now that was easy, wasn't it?

2)

Dropped the dependency on actionmailer since
there's really no need for it (This will most
probably need updates to

  http://github.com/snusnu/rails-templates

though, because a default generated rails app gets
generated with config values that rely on the
presence of actionmailer.

3)

We also don't require activemodel explicitly
anymore. This is because active_model already
does that for us. No need specifying dependencies
twice.

Thx to Alex Mankuta and Dan Kubb for noticing
the dependency weirdness.

4)

Added instructions on how to get a development
environment up and running to the README